Slight refactor of set_to_array
Have it take a path_id instead of a Set, and move it to relctx.  (I
think I might need to call it from unpack_rvar, but it seems a
reasonable change anyway.)
